place test in correct file
